packages/backend/pyproject.toml
[tool.poetry]
name = "proxy-accounting-backend"
version = "0.1.0"
description = "代理记账营运内部系统后端"
authors = ["Developer <dev@example.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.scripts]
dev = "uvicorn src.main:app --reload --port 8000"
start = "uvicorn src.main:app --host 0.0.0.0 --port 8000"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.1"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
sqlalchemy = "^2.0.23"
psycopg2-binary = "^2.9.9"
alembic = "^1.13.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.6"
pydantic-settings = "^2.1.0"
pyjwt = "^2.10.1"
email-validator = "^2.3.0"
redis = "^5.0.0"
# 支付相关依赖
wechatpayv3 = "^1.2.6"
alipay-sdk-python = "^3.7.4"
cryptography = "^41.0.7"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
httpx = "^0.25.2"
black = "^23.11.0"
isort = "^5.12.0"
flake8 = "^6.1.0"
mypy = "^1.7.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py311']

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
